Enum lldb::StateType

pub enum StateType {
    Invalid,
    Unloaded,
    Connected,
    Attaching,
    Launching,
    Stopped,
    Running,
    Stepping,
    Crashed,
    Detached,
    Exited,
    Suspended,
}

Process and thread states.

Variants

Invalid

Unloaded

Process object is valid, but not currently loaded.

Connected

Process is connected to remote debug services, but not launched or attached to anything yet.

Attaching

Process is currently trying to attach.

Launching

Process is currently trying to launch.

Stopped

Process or thread is stopped and can be examined.

Running

Process or thread is running and can't be examined.

Stepping

Process or thread is in the process of stepping and can't be examined.

Crashed

Process or thread has crashed and can be examined.

Detached

Process has been detached and can't be examined.

Exited

Process has exited and can't be examined.

Suspended

Process or thread is in a suspended state as far as the debugger is concerned while other processes or threads get the chance to run.