Enum lldb::SymbolContextItem

pub enum SymbolContextItem {
    Target,
    Module,
    CompUnit,
    Function,
    Block,
    LineEntry,
    Symbol,
    Everything,
    Variable,
}

These mask bits allow a common interface for queries that can limit the amount of information that gets parsed to only the information that is requested. These bits also can indicate what actually did get resolved during query function calls.

Each definition corresponds to a one of the member variables in this class, and requests that that item be resolved, or indicates that the member did get resolved.

Variants

Target

Set when a target is requested from a query or was located in the query results.

Module

Set when a module is requested from a query or was located in the query results.

CompUnit

Set when a compilation unit is requested from a query or was located in the query results.

Function

Set when a function is requested from a query or was located in the query results.

Block

Set when the deepest block is requested from a query or was located in the query results.

LineEntry

Set when a line entry is requested from a query or was located in the query results.

Symbol

Set when a symbol is requested from a query or was located in the query results.

Everything

Indicates to try and look everything up during a routine symbol context query. This doesn't actually include looking up a variable.

Variable

Set when a global or static variable is requested from a query, or was located in the query results.

This is potentially expensive to look up, so it isn't included in Everything which stops it from being used during frame PC lookups and many other potential address to symbol context lookups.