Struct lldb::LaunchFlags

#[repr(C)]

pub struct LaunchFlags { /* fields omitted */ }

Methods

impl LaunchFlags

const EXEC: LaunchFlags

EXEC: LaunchFlags = LaunchFlags{bits: 1,}

const DEBUG: LaunchFlags

DEBUG: LaunchFlags = LaunchFlags{bits: 2,}

const STOP_AT_ENTRY: LaunchFlags

STOP_AT_ENTRY: LaunchFlags = LaunchFlags{bits: 4,}

const DISABLE_ASLR: LaunchFlags

DISABLE_ASLR: LaunchFlags = LaunchFlags{bits: 8,}

const DISABLE_STDIO: LaunchFlags

DISABLE_STDIO: LaunchFlags = LaunchFlags{bits: 16,}

const LAUNCH_IN_TTY: LaunchFlags

LAUNCH_IN_TTY: LaunchFlags = LaunchFlags{bits: 32,}

const LAUNCH_IN_SHELL: LaunchFlags

LAUNCH_IN_SHELL: LaunchFlags = LaunchFlags{bits: 64,}

const LAUNCH_IN_SEPARATE_PROCESS_GROUP: LaunchFlags

LAUNCH_IN_SEPARATE_PROCESS_GROUP: LaunchFlags = LaunchFlags{bits: 128,}

const DONT_SET_EXIT_STATUS: LaunchFlags

DONT_SET_EXIT_STATUS: LaunchFlags = LaunchFlags{bits: 256,}

const DETACH_ON_ERRROR: LaunchFlags

DETACH_ON_ERRROR: LaunchFlags = LaunchFlags{bits: 512,}

const SHELL_EXPAND_ARGUMENTS: LaunchFlags

SHELL_EXPAND_ARGUMENTS: LaunchFlags = LaunchFlags{bits: 1024,}

const CLOSE_TTY_ON_EXIT: LaunchFlags

CLOSE_TTY_ON_EXIT: LaunchFlags = LaunchFlags{bits: 2048,}

fn empty() -> LaunchFlags

Returns an empty set of flags.

fn all() -> LaunchFlags

Returns the set containing all flags.

fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

fn from_bits(bits: u32) -> Option<LaunchFlags>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: u32) -> LaunchFlags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Returns true if all flags are currently set.

fn intersects(&self, other: LaunchFlags) -> bool

Returns true if there are flags common to both self and other.

fn contains(&self, other: LaunchFlags) -> bool

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: LaunchFlags)

Inserts the specified flags in-place.

fn remove(&mut self, other: LaunchFlags)

Removes the specified flags in-place.

fn toggle(&mut self, other: LaunchFlags)

Toggles the specified flags in-place.

fn set(&mut self, other: LaunchFlags, value: bool)

Inserts or removes the specified flags depending on the passed value.