Struct lldb::SBAttachInfo

pub struct SBAttachInfo {
    pub raw: SBAttachInfoRef,
}

Configuration for attaching to a process.

See SBTarget::attach.

Fields
raw: SBAttachInfoRef

The underlying raw SBAttachInfoRef.

Methods

impl SBAttachInfo

fn new() -> SBAttachInfo

Construct a new SBAttachInfo.

fn new_with_pid(pid: lldb_pid_t) -> SBAttachInfo

Construct a new SBAttachInfo for a given process ID (pid).

fn new_with_path(path: &str, wait_for: bool, async: bool) -> SBAttachInfo

Attach to a process by name.

Future calls to SBTarget::attach(...) will be synchronous or asynchronous depending on the async argument.

path: A full or partial name for the process to attach to.

wait_for: If false, attach to an existing process whose name matches. If true, then wait for the next process whose name matches.

async: If false, then the SBTarget::attach call will be synchronous with no way to cancel the attach while it is in progress. If true, then the SBTarget::attach call will return immediately and clients are expected to wait for a process eStateStopped event if a suitable process is eventually found. If the client wants to cancel the event, SBProcess::stop can be called and an eStateExited process event will be delivered.

fn wrap(raw: SBAttachInfoRef) -> SBAttachInfo

Construct a new SBAttachInfo.

fn process_id(&self) -> lldb_pid_t

fn set_process_id(&self, pid: lldb_pid_t)

fn set_executable_path(&self, path: &str)

fn set_executable_filespec(&self, exe_file: SBFileSpec)

fn ignore_existing(&self) -> bool

fn set_ignore_existing(&self, b: bool)

fn resume_count(&self) -> u32

fn set_resume_count(&self, c: u32)

fn listener(&self) -> SBListener

Get the listener that will be used to receive process events.

If no listener has been set via a call to SBAttachInfo::set_listener(), then an invalid SBListener will be returned (SBListener::is_valid() will return false). If a listener has been set, then the valid listener object will be returned.

fn set_listener(&self, listener: SBListener)

Set the listener that will be used to receive process events.

By default the SBDebugger, which has a listener, that the SBTarget belongs to will listen for the process events. Calling this function allows a different listener to be used to listen for process events.