Struct lldb::SBBlock

pub struct SBBlock {
    pub raw: SBBlockRef,
}

A lexical block.

Fields

raw

The underlying raw SBBlockRef.

Methods

impl SBBlock

fn wrap(raw: SBBlockRef) -> SBBlock

Construct a new SBBlock.

fn maybe_wrap(raw: SBBlockRef) -> Option<SBBlock>

Construct a new Some(SBBlock) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBBlock value.

fn is_inlined(&self) -> bool

Does this block represent an inlined function?

fn inlined_name(&self) -> &str

Get the function name if this block represents an inlined function.

fn inlined_call_site_file(&self) -> Option<SBFileSpec>

Get the call site file if this block represents an inlined function.

fn inlined_call_site_line(&self) -> Option<u32>

Get the call site line number if this block represents an inlined function.

fn inlined_call_site_column(&self) -> Option<u32>

Get the call site column number if this block represents an inlined function.

fn parent(&self) -> Option<SBBlock>

Get the parent block

fn containing_inlined_block(&self) -> Option<SBBlock>

Get the inlined block that is or contains this block.

fn sibling(&self) -> Option<SBBlock>

Get the sibling block for this block.

fn first_child(&self) -> Option<SBBlock>

Get the first child block for this block.

fn num_ranges(&self) -> u32

The number of address ranges associated with this block.

fn range_start_address(&self, idx: u32) -> SBAddress

Get the start address of an address range.

fn range_end_address(&self, idx: u32) -> SBAddress

Get the end address of an address range.

fn range_index_for_block_address(&self, block_address: &SBAddress) -> u32

Given an address, find out which address range it is part of.