Struct lldb::SBDebugger

pub struct SBDebugger {
    pub raw: SBDebuggerRef,
}

Creates SBTargets, provides access to them and manages the overall debugging experience.

Fields

raw

The underlying raw SBDebuggerRef.

Methods

impl SBDebugger

fn initialize()

Initialize LLDB.

This should be called before LLDB functionality is used.

fn terminate()

Tear down LLDB.

This should be called once the application no longer needs to use LLDB functionality. Typically, this is called as the application exits.

fn create(source_init_files: bool) -> SBDebugger

Create a new instance of SBDebugger.

If source_init_files is true, then ~/.lldbinit will be processed.

fn async(&self) -> bool

Get whether or not the debugger is in async mode.

When in async mode, the debugger returns immediately when stepping or continuing without waiting for the process to change state.

fn set_async(&mut self, async: bool)

Set the debugger to be in async mode or not.

When in async mode, the debugger returns immediately when stepping or continuing without waiting for the process to change state.

fn version() -> String

Get the LLDB version string.

fn targets(&self) -> DebuggerTargetIter

Get an iterator over the targets known to this debugger instance.

fn selected_platform(&self) -> SBPlatform

Get the currently selected SBPlatform.

fn set_selected_platform(&mut self, platform: &SBPlatform)

Set the selected SBPlatform.