Struct lldb::SBError

pub struct SBError {
    pub raw: SBErrorRef,
}

A container for holding any error code.

Fields
raw: SBErrorRef

The underlying raw SBErrorRef.

Methods

impl SBError

fn new() -> SBError

Construct a new SBError.

fn wrap(raw: SBErrorRef) -> SBError

Construct a new SBError.

fn maybe_wrap(raw: SBErrorRef) -> Option<SBError>

Construct a new Some(SBError) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBError value.

fn error_string(&self) -> &str

Any textual error message associated with the error.

fn is_failure(&self) -> bool

Does this error represent a failure?

fn is_success(&self) -> bool

Does this error represent a success?

fn error(&self) -> u32

The underlying error code. Must be interpreted in conjunction with the error type.

fn error_type(&self) -> ErrorType

What type of error is this?