Struct lldb::SBLaunchInfo

pub struct SBLaunchInfo {
    pub raw: SBLaunchInfoRef,
}

Configuration for launching a process.

See SBTarget::launch.

Fields
raw: SBLaunchInfoRef

The underlying raw SBLaunchInfoRef.

Methods

impl SBLaunchInfo

fn new() -> SBLaunchInfo

Construct a new SBLaunchInfo.

fn wrap(raw: SBLaunchInfoRef) -> SBLaunchInfo

Construct a new SBLaunchInfo.

fn process_id(&self) -> lldb_pid_t

fn user_id(&self) -> u32

fn set_user_id(&self, user_id: u32)

fn user_id_is_valid(&self) -> bool

fn group_id(&self) -> u32

fn set_group_id(&self, group_id: u32)

fn group_id_is_valid(&self) -> bool

fn executable_file(&self) -> Option<SBFileSpec>

fn set_executable_file(&self, filespec: &SBFileSpec, add_as_first_arg: bool)

Set the executable file that will be used to launch the process and optionally set it as the first argument in the argument vector.

This only needs to be specified if clients wish to carefully control the exact path will be used to launch a binary. If you create a target with a symlink, that symlink will get resolved in the target and the resolved path will get used to launch the process. Calling this function can help you still launch your process using the path of your choice.

If this function is not called prior to launching with SBTarget::launch(...), the target will use the resolved executable path that was used to create the target.

exe_file is the override path to use when launching the executable.

If add_as_first_arg is true, then the path will be inserted into the argument vector prior to launching. Otherwise the argument vector will be left alone.

fn listener(&self) -> Option<SBListener>

Get the listener that will be used to receive process events.

If no listener has been set via a call to SBLaunchInfo::set_listener(), then None will be returned. If a listener has been set, then the listener object will be returned.

fn set_listener(&self, listener: &SBListener)

Set the listener that will be used to receive process events.

By default the SBDebugger, which has a listener, that the SBTarget belongs to will listen for the process events. Calling this function allows a different listener to be used to listen for process events.

fn launch_flags(&self) -> LaunchFlags

fn set_launch_flags(&self, launch_flags: LaunchFlags)

fn process_plugin_name(&self) -> Option<&str>

fn set_process_plugin_name(&self, plugin: &str)

fn shell(&self) -> Option<&str>

fn set_shell(&self, shell: &str)

fn shell_expand_arguments(&self) -> bool

fn set_shell_expand_arguments(&self, expand: bool)

fn resume_count(&self) -> u32

fn set_resume_count(&self, resume_count: u32)

fn add_close_file_action(&self, fd: i32) -> bool

fn add_duplicate_file_action(&self, fd: i32, dup_fd: i32) -> bool

fn add_open_file_action(
&self,
fd: i32,
path: &str,
read: bool,
write: bool
) -> bool

fn add_suppress_file_action(&self, fd: i32, read: bool, write: bool) -> bool

fn launch_event_data(&self) -> Option<&str>

fn set_launch_event_data(&self, data: &str)

fn detach_on_error(&self) -> bool

fn set_detach_on_error(&self, detach: bool)