Struct lldb::SBLineEntry

pub struct SBLineEntry {
    pub raw: SBLineEntryRef,
}

Specifies an association with a contiguous range of instructions and a source file location.

Fields
raw: SBLineEntryRef

The underlying raw SBLineEntryRef.

Methods

impl SBLineEntry

fn wrap(raw: SBLineEntryRef) -> SBLineEntry

Construct a new SBLineEntry.

fn maybe_wrap(raw: SBLineEntryRef) -> Option<SBLineEntry>

Construct a new Some(SBLineEntry) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBLineEntry value.

fn start_address(&self) -> SBAddress

The start address for this line entry.

fn end_address(&self) -> SBAddress

The end address for this line entry.

fn filespec(&self) -> SBFileSpec

The file (SBFileSpec) for this line entry.

fn line(&self) -> u32

The 1-based line number for this line entry.

A return value of 0 indicates that no line information is available.

fn column(&self) -> u32

The 1-based column number for this line entry.

A return value of 0 indicates that no column information is available.