Struct lldb::SBModule

pub struct SBModule {
    pub raw: SBModuleRef,
}

An executable image and its associated object and symbol files.

Fields
raw: SBModuleRef

The underlying raw SBModuleRef.

Methods

impl SBModule

fn wrap(raw: SBModuleRef) -> SBModule

Construct a new SBModule.

fn maybe_wrap(raw: SBModuleRef) -> Option<SBModule>

Construct a new Some(SBModule) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBModule value.

fn filespec(&self) -> SBFileSpec

The file for the module on the host system that is running LLDB.

This can differ from the path on the platform since we might be doing remote debugging.

fn platform_filespec(&self) -> SBFileSpec

The file for the module as it is known on the remote system on which it is being debugged.

For local debugging this is always the same as SBModule::filespec. But remote debugging might mention a file /usr/lib/liba.dylib which might be locally downloaded and cached. In this case the platform file could be something like: /tmp/lldb/platform-cache/remote.host.computer/usr/lib/liba.dylib The file could also be cached in a local developer kit directory.

fn find_section(&self, name: &str) -> Option<SBSection>

fn sections(&self) -> SBModuleSectionIter

Get an iterator over the sections known to this module instance.