Struct lldb::SBModuleSpec

pub struct SBModuleSpec {
    pub raw: SBModuleSpecRef,
}

A description of an SBModule.

Fields
raw: SBModuleSpecRef

The underlying raw SBModuleSpecRef.

Methods

impl SBModuleSpec

fn wrap(raw: SBModuleSpecRef) -> SBModuleSpec

Construct a new SBModuleSpec.

fn maybe_wrap(raw: SBModuleSpecRef) -> Option<SBModuleSpec>

Construct a new Some(SBModuleSpec) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBModuleSpec value.

fn filespec(&self) -> SBFileSpec

The file for the module on the host system that is running LLDB.

This can differ from the path on the platform since we might be doing remote debugging.

fn set_filespec(&self, filespec: &SBFileSpec)

Set the file for the module on the host system that is running LLDB.

fn platform_filespec(&self) -> SBFileSpec

The file for the module as it is known on the remote system which is being debugged.

For local debugging, this is always the same as SBModuleSpec::filespec. But remote debugging might mention a file /usr/lib/liba.dylib which might be locally downloaded and cached. In this case, the platform file could be something like: /tmp/lldb/platform-cache/remote.host.computer/usr/lib/liba.dylib The file could also be cached in a local developer kit directory.

fn set_platform_filespec(&self, filespec: &SBFileSpec)

Set the file for the module as it is known on the remote system which is being debugged.

fn symbol_filespec(&self) -> Option<SBFileSpec>

fn set_symbol_filespec(&self, filespec: &SBFileSpec)

fn object_name(&self) -> &str

fn set_object_name(&self, _object_name: &str)

fn triple(&self) -> &str

fn set_triple(&self, _object_name: &str)

fn uuid_bytes(&self) -> &str

fn set_uuid_bytes(&self, _object_name: &str)