Struct lldb::SBPlatform

pub struct SBPlatform {
    pub raw: SBPlatformRef,
}

A platform that can represent the current host or a remote host debug platform.

The SBPlatform class represents the current host, or a remote host. It can be connected to a remote platform in order to provide ways to remotely launch and attach to processes, upload/download files, create directories, run remote shell commands, find locally cached versions of files from the remote system, and much more.

SBPlatform objects can be created and then used to connect to a remote platform which allows the SBPlatform to be used to get a list of the current processes on the remote host, attach to one of those processes, install programs on the remote system, attach and launch processes, and much more.

Every SBTarget has a corresponding SBPlatform. The platform can be specified upon target creation, or the currently selected platform will attempt to be used when creating the target automatically as long as the currently selected platform matches the target architecture and executable type. If the architecture or executable type do not match, a suitable platform will be found automatically.

Fields
raw: SBPlatformRef

The underlying raw SBPlatformRef.

Methods

impl SBPlatform

fn wrap(raw: SBPlatformRef) -> SBPlatform

Construct a new SBPlatform.

fn maybe_wrap(raw: SBPlatformRef) -> Option<SBPlatform>

Construct a new Some(SBPlatform) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBPlatform value.

fn working_directory(&self) -> &str

The working directory for this platform.

fn name(&self) -> &str

The name of the platform.

When debugging on the host platform, this would be "host".

fn triple(&self) -> &str

The triple used to describe this platform.

An example value might be "x86_64-apple-macosx".

fn hostname(&self) -> &str

The hostname for this platform.

fn os_build(&self) -> &str

The build ID for the platforms' OS version.

fn os_description(&self) -> &str

The long form description of the platform's OS version.

On Mac OS X, this might look like "Darwin Kernel Version 15.5.0: Tue Apr 19 18:36:36 PDT 2016; root:xnu-3248.50.21~8/RELEASE_X86_64".

fn os_major_version(&self) -> u32

The major component of the platform's OS version.

On Mac OS X 10.11.4, this would have the value 10.

fn os_minor_version(&self) -> u32

The minor component of the platform's OS version.

On Mac OS X 10.11.4, this would have the value 11.

fn os_update_version(&self) -> u32

The patch or update component of the platform's OS version.

On Mac OS X 10.11.4, this would have the value 4.

fn launch(&self, launch_info: &SBLaunchInfo) -> Result<(), SBError>

Launch a process. This is not for debugging that process.

fn kill(&self, pid: lldb_pid_t) -> Result<(), SBError>

Kill a process.