Struct lldb::SBQueueItem

pub struct SBQueueItem {
    pub raw: SBQueueItemRef,
}

A work item enqueued on a libdispatch aka Grand Central Dispatch (GCD) queue.

Most often, this will be a function or block.

"enqueued" here means that the work item has been added to a queue but it has not yet started executing. When it is "dequeued", execution of the item begins.

Fields
raw: SBQueueItemRef

The underlying raw SBQueueItemRef.

Methods

impl SBQueueItem

fn wrap(raw: SBQueueItemRef) -> SBQueueItem

Construct a new SBQueueItem.

fn maybe_wrap(raw: SBQueueItemRef) -> Option<SBQueueItem>

Construct a new Some(SBQueueItem) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBQueueItem value.

fn kind(&self) -> QueueItemKind

The kind of this work item.

fn address(&self) -> Option<SBAddress>

The code address that will be executed when this work item is executed.

Not all queue items will have an address associated with them. QueueItemKind::Function and QueueItemKind::Block work items should have an address.

fn extended_backtrace_thread(&self, thread_type: &str) -> Option<SBThread>

Get an extended backtrace thread for this queue item, if available

If the backtrace/thread information was collected when this item was enqueued, this call will provide it.

The thread_type will typically be one of "libdispatch" or "pthread".