Struct lldb::SBSection

pub struct SBSection {
    pub raw: SBSectionRef,
}

Fields
raw: SBSectionRef

The underlying raw SBSectionRef.

Methods

impl SBSection

fn wrap(raw: SBSectionRef) -> SBSection

Construct a new SBSection.

fn maybe_wrap(raw: SBSectionRef) -> Option<SBSection>

Construct a new Some(SBSection) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBSection value.

fn name(&self) -> &str

The section name.

fn parent(&self) -> Option<SBSection>

The section parent, if there is one.

fn find_subsection(&self, name: &str) -> Option<SBSection>

fn subsections(&self) -> SBSectionSubSectionIter

Get an iterator over the subsections known to this section instance.

fn file_address(&self) -> u64

fn load_address(&self, target: &SBTarget) -> u64

fn byte_size(&self) -> u64

fn file_offset(&self) -> u64

fn file_byte_size(&self) -> u64

fn section_data(&self) -> SBData

fn section_data_slice(&self, offset: u64, size: u64) -> SBData

fn section_type(&self) -> SectionType

fn target_byte_size(&self) -> u32