Struct lldb::SBStructuredData

pub struct SBStructuredData {
    pub raw: SBStructuredDataRef,
}

The value of a variable, register or expression.

Fields
raw: SBStructuredDataRef

The underlying raw SBStructuredDataRef.

Methods

impl SBStructuredData

fn wrap(raw: SBStructuredDataRef) -> SBStructuredData

Construct a new SBStructuredData.

fn maybe_wrap(raw: SBStructuredDataRef) -> Option<SBStructuredData>

Construct a new Some(SBStructuredData) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBStructuredData value.

fn clear(&self)

fn set_from_json(&self, stream: &SBStream) -> Result<(), SBError>

fn get_as_json(&self) -> Result<SBStream, SBError>

fn data_type(&self) -> StructuredDataType

Return the type of data in this data structure.

fn size(&self) -> usize

Return the size (number of elements) in this data structure if it is an array or dictionary type. For other types, 0 will be returned.

fn value_for_key(&self, key: &str) -> Option<SBStructuredData>

Return the value corresponding to a key if this data structure is a dictionary type.

fn item_at_index(&self, idx: usize) -> Option<SBStructuredData>

Return the value corresponding to an index if this data structure is array.

fn integer_value(&self) -> Option<u64>

Return the integer value if this data structure is an integer type.

fn float_value(&self) -> Option<f64>

Return the floating point value if this data structure is a floating type.

fn boolean_value(&self) -> Option<bool>

Return the boolean value if this data structure is a boolean type.

fn string_value(&self) -> Option<String>

Provides the string value if this data structure is a string type.