Struct lldb::SBValue

pub struct SBValue {
    pub raw: SBValueRef,
}

The value of a variable, register or expression.

Fields
raw: SBValueRef

The underlying raw SBValueRef.

Methods

impl SBValue

fn wrap(raw: SBValueRef) -> SBValue

Construct a new SBValue.

fn maybe_wrap(raw: SBValueRef) -> Option<SBValue>

Construct a new Some(SBValue) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBValue value.

fn dereference(&self) -> Option<SBValue>

fn address_of(&self) -> Option<SBValue>

fn type_is_pointer_type(&self) -> bool

fn target(&self) -> SBTarget

fn process(&self) -> SBProcess

fn thread(&self) -> SBThread

fn frame(&self) -> SBFrame

fn watch(&self, resolve_location: bool, read: bool, write: bool) -> Result<SBWatchpoint, SBError>

Find and watch a variable.

fn watch_pointee(&self, resolve_location: bool, read: bool, write: bool) -> Result<SBWatchpoint, SBError>

Find and watch the location pointed to by a variable.

fn pointee_data(&self, item_idx: u32, item_count: u32) -> Option<SBData>

Get an SBData wrapping what this SBValue points to.

This method will dereference the current SBValue, if its data type is a T* or T[], and extract item_count elements of type T from it, copying their contents into an SBData.

item_idx is the index of the first item to retrieve. For an array this is equivalent to array[item_idx], for a pointer to *(pointer + item_idx). In either case, the measurement unit for item_idx is the sizeof(T) rather than bytes.

item_count is how many items should be copied into the output. By default only one item is copied, but more can be asked for.

Returns Some(SBData) with the contents of the copied items, on success. None otherwise.

fn data(&self) -> Option<SBData>

Get an SBData wrapping the contents of this SBValue.

This method will read the contents of this object in memory and copy them into an SBData for future use.

Returns Some(SBData) with the contents of this SBValue, on success. None otherwise.

fn set_data(&self, data: &SBData) -> Result<(), SBError>

fn load_address(&self) -> Option<lldb_addr_t>

fn address(&self) -> Option<SBAddress>