Struct lldb::SBWatchpoint

pub struct SBWatchpoint {
    pub raw: SBWatchpointRef,
}

An instance of a watch point for a specific target program.

A watchpoint is determined by the address the byte size that resulted in this particular instantiation. Each watchpoint has its own settable options.

To Hit or Not

A watchpoint has multiple ways of controlling whether or not it should be considered active.

Enabled. This is controlled via is_enabled and set_enabled.

Ignore count. If set, this watchpoint will be ignored the first ignore count times that it is hit. This is controlled via ignore_count and set_ignore_count.

A count of how many times a watchpoint has been it is available via hit_count.

Fields
raw: SBWatchpointRef

The underlying raw SBWatchpointRef.

Methods

impl SBWatchpoint

fn wrap(raw: SBWatchpointRef) -> SBWatchpoint

Construct a new SBWatchpoint.

fn maybe_wrap(raw: SBWatchpointRef) -> Option<SBWatchpoint>

Construct a new Some(SBWatchpoint) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBWatchpoint value.

fn error(&self) -> Option<SBError>

fn id(&self) -> i32

fn hardware_index(&self) -> Option<i32>

fn watch_address(&self) -> lldb_addr_t

fn watch_size(&self) -> u32

fn is_enabled(&self) -> bool

fn set_enabled(&self, enabled: bool)

fn hit_count(&self) -> u32

fn ignore_count(&self) -> u32

fn set_ignore_count(&self, count: u32)