Crate lldb

LLDB

This crate provides a safe binding to the public API for LLDB, the debugger provided by the LLVM project. LLDB provides a modern, high performance debugger framework and is the default debugger for Mac OS X and iOS.

Installation

This crate works with Cargo and is on crates.io. Add it to your Cargo.toml like so:

lldb = "0.0.2"

Mac OS X Installation Notes

On Mac OS X, this library relies upon being able to find the LLDB.framework that is provided by Xcode.app.

In your own application, you will need to configure the @rpath for your execuable to point to the location of the LLDB.framework that you are using. Doing this automatically is not currently supported by Cargo.

For testing, you can use the version provided by Xcode.app by setting an environment variable:

export DYLD_FRAMEWORK_PATH=/Applications/Xcode.app/Contents/SharedFrameworks

Linux Installation Notes

Support for building this has not yet been provided for Linux. Contributions are welcome!

Windows Installation Notes

Support for building this has not yet been provided for Linux. Contributions are welcome!

Usage

The primary entry point is SBDebugger. This will be how you create a debug target and begin the actually interesting stuff.

Important Classes

The LLDB API provides many structs and a wide range of functionality. Some of the most common usages will involve these structs and their corresponding methods:

SBDebugger: Manages the entire debug experience and creates SBTargets.

SBTarget: The target program running under the debugger.

SBProcess: The process associated with the target program.

SBThread: A thread of execution. SBProcess contains SBThreads.

SBFrame: One of the stack frames associated with a thread. SBThread contains SBFrames.

SBSymbolContext: A container that stores various debugger related info.

SBValue: The value of a variable, a register, or an expression.

SBModule: An executable image and its associated object and symbol files. SBTarget contains SBModules.

SBBreakpoint: A logical breakpoint and its associated settings. SBTarget contains SBBreakpoints.

SBSymbol: The symbol possibly associated with a stack frame.

SBCompileUnit: A compilation unit, or compiled source file.

SBFunction: A generic function, which can be inlined or not.

SBBlock: A lexical block. SBFunction contains SBBlocks.

SBLineEntry: Specifies an association with a contiguous range of instructions and a source file location. SBCompileUnit contains SBLineEntrys.

Development Guidelines

The official LLDB bindings for C++ and Python maintain very strict backwards compatibility. This has resulted in them having methods that are considered to be deprecated in favor of newer methods with more comprehensive arguments. The Rust bindings don't have that problem (yet), so we should try to only write bindings for the recommended usages and not the deprecated methods.

The documentation for the official bindings is also somewhat spotty. While it would be great to contribute upstream to improve that, we should seek to have these bindings be well documented and with more examples.

People should be jealous of the quality of our bindings. Sic itur ad astra.

Structs

SBAddress
A section + offset based address class.

SBAttachInfo
Configuration for attaching to a process.

SBBlock
A lexical block.

SBBreakpoint
A logical breakpoint and its associated settings.

SBBroadcaster
An entity which can broadcast events.

SBCompileUnit
A compilation unit or compiled source file.

SBData
A block of data.

SBDebugger
Creates SBTargets, provides access to them and manages the overall debugging experience.

SBError
A container for holding any error code.

SBEvent
An event.

SBFileSpec
A file specification that divides the path into a directory and basename.

SBFrame
One of the stack frames associated with a thread.

SBFunction
A generic function, which can be inlined or not.

SBInstruction
A machine instruction.

SBInstructionList
A list of machine instructions.

SBLaunchInfo
Configuration for launching a process.

SBLineEntry
Specifies an association with a contiguous range of instructions and a source file location.

SBListener
Listen for debugger events.

SBModule
An executable image and its associated object and symbol files.

SBModuleSpec
A description of an SBModule.

SBPlatform
A platform that can represent the current host or a remote host debug platform.

SBProcess
The process associated with the target program.

SBSymbol
The symbol possibly associated with a stack frame.

SBSymbolContext
A container that stores various debugger related info.

SBTarget
The target program running under the debugger.

SBThread
A thread of execution.

SBValue
The value of a variable, register or expression.

Enums

AccessType

AddressClass

BasicType

BreakpointEventType

ByteOrder
Byte order definitions.

CommandArgumentType

CommandFlags

ConnectionStatus

DescriptionLevel

DisassemblyFlavor
Which syntax should be used in disassembly?

DynamicValueType

EmulateInstructionOptions

Encoding
Register encoding defintions.

ErrorType

ExpressionEvaluationPhase

ExpressionResults

FilePermissions

Format
Display format definitions.

FrameComparison

FunctionNameType

GdbSignal

InputReaderAction

InputReaderGranularity

InstrumentationRuntimeType

LanguageType

LaunchFlags

MatchType

MemberFunctionKind

PathType

Permissions

QueueItemKind

QueueKind

RegisterKind

ReturnStatus

RunMode
Thread run modes.

ScriptLanguage

SectionType

StateType
Process and thread states.

StopReason
Thread stop reasons.

SymbolContextItem
These mask bits allow a common interface for queries that can limit the amount of information that gets parsed to only the information that is requested. These bits also can indicate what actually did get resolved during query function calls.

SymbolType

TemplateArgumentKind

TypeClass

TypeFlags

TypeOptions

TypeSummaryCapping

ValueType

WatchpointEventType

WatchpointKind

Type Definitions

lldb_addr_t

lldb_offset_t

lldb_pid_t

lldb_tid_t

lldb_user_id_t