Crate lldb

LLDB

Installation

This crate works with Cargo and is on crates.io. Add it to your Cargo.toml like so:

lldb = "0.0.1"

Mac OS X Installation Notes

On Mac OS X, this library relies upon being able to find the LLDB.framework that is provided by Xcode.app.

In your own application, you will need to configure the @rpath for your execuable to point to the location of the LLDB.framework that you are using. Doing this automatically is not currently supported by Cargo.

For testing, you can use the version provided by Xcode.app by setting an environment variable:

export DYLD_FRAMEWORK_PATH=/Applications/Xcode.app/Contents/SharedFrameworks

Usage

LLDB must be initialized before the functionality is used. This is done with SBDebugger::initialize():

use lldb;

lldb::SBDebugger::initialize();

Similarly, it must be terminated after you are done using it:

use lldb;

lldb::SBDebugger::initialize();
// Use LLDB functionality ...
lldb::SBDebugger::terminate();

The primary entry point from here will be SBDebugger. This will be how you create a debug target and begin the actually interesting stuff.

Important Classes

The LLDB API provides many structs and a wide range of functionality. Some of the most common usages will involve these structs and their corresponding methods:

SBDebugger: Manages the entire debug experience and creates SBTargets.

SBTarget: The target program running under the debugger.

SBProcess: The process associated with the target program.

SBThread: A thread of execution. SBProcess contains SBThreads.

SBFrame: One of the stack frames associated with a thread. SBThread contains SBFrames.

SBSymbolContext: A container that stores various debugger related info.

SBValue: The value of a variable, a register, or an expression.

SBModule: An executable image and its associated object and symbol files. SBTarget contains SBModules.

SBBreakpoint: A logical breakpoint and its associated settings. SBTarget contains SBBreakpoints.

SBSymbol: The symbol possibly associated with a stack frame.

SBCompileUnit: A compilation unit, or compiled source file.

SBFunction: A generic function, which can be inlined or not.

SBBlock: A lexical block. SBFunction contains SBBlocks.

SBLineEntry: Specifies an association with a contiguous range of instructions and a source file location. SBCompileUnit contains SBLineEntrys.

Development Guidelines

The official LLDB bindings for C++ and Python maintain very strict backwards compatibility. This has resulted in them having methods that are considered to be deprecated in favor of newer methods with more comprehensive arguments. The Rust bindings don't have that problem (yet), so we should try to only write bindings for the recommended usages and not the deprecated methods.

The documentation for the official bindings is also somewhat spotty. While it would be great to contribute upstream to improve that, we should seek to have these bindings be well documented and with more examples.

People should be jealous of the quality of our bindings. Sic itur ad astra.

Structs

SBAddress
A section + offset based address class.

SBBlock
A lexical block.

SBBreakpoint
A logical breakpoint and its associated settings.

SBCompileUnit
A compilation unit or compiled source file.

SBData
A block of data.

SBDebugger
Creates SBTargets, provides access to them and manages the overall debugging experience.

SBFileSpec
A file specification that divides the path into a directory and basename.

SBFrame
One of the stack frames associated with a thread.

SBFunction
A generic function, which can be inlined or not.

SBInstruction
A machine instruction.

SBInstructionList
A list of machine instructions.

SBLineEntry
Specifies an association with a contiguous range of instructions and a source file location.

SBModule
An executable image and its associated object and symbol files.

SBModuleSpec
A description of an SBModule.

SBPlatform
A platform that can represent the current host or a remote host debug platform.

SBProcess
The process associated with the target program.

SBSymbol
The symbol possibly associated with a stack frame.

SBSymbolContext
A container that stores various debugger related info.

SBTarget
The target program running under the debugger.

SBThread
A thread of execution.

SBValue
The value of a variable, register or expression.