Struct lldb::SBBreakpoint

pub struct SBBreakpoint {
    pub raw: SBBreakpointRef,
}

A logical breakpoint and its associated settings.

To Hit or Not

A breakpoint has multiple ways of controlling whether or not it should be considered active.

Enabled. This is controlled via is_enabled and set_enabled.

One shot. If set, this will be disabled once it has been hit. This is controlled via is_oneshot and set_oneshot.

Ignore count. If set, this breakpoint will be ignored the first ignore count times that it is hit. This is controlled via ignore_count and set_ignore_count.

A count of how many times a breakpoint has been it is available via hit_count.

Breakpoint Names and Aliases

Breakpoints can have names associated with them. These are actually more like tags in that the same name can be applied to multiple breakpoints so that a single command invocation can work on multiple breakpoints at once.

A common use case for this is setting up families of breakpoints, for example on malloc, realloc, and free and giving them all a name of memory. Then, you can make it easy for the user enable or disable them all in a single shot.

Names are managed via add_name, remove_name, matches_name and names.

Breakpoint Locations

...

Fields
raw: SBBreakpointRef

The underlying raw SBBreakpointRef.

Methods

impl SBBreakpoint

fn wrap(raw: SBBreakpointRef) -> SBBreakpoint

Construct a new SBBreakpoint.

fn maybe_wrap(raw: SBBreakpointRef) -> Option<SBBreakpoint>

Construct a new Some(SBBreakpoint) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBBreakpoint value.

fn id(&self) -> i32

fn is_enabled(&self) -> bool

fn set_enabled(&self, enabled: bool)

fn is_oneshot(&self) -> bool

fn set_oneshot(&self, oneshot: bool)

fn is_internal(&self) -> bool

fn hit_count(&self) -> u32

fn ignore_count(&self) -> u32

fn set_ignore_count(&self, count: u32)

fn add_name(&self, name: &str) -> bool

fn remove_name(&self, name: &str)

fn matches_name(&self, name: &str) -> bool

fn names(&self) -> SBStringList

fn clear_all_breakpoint_sites(&self)

fn find_location_by_address(
&self,
address: lldb_addr_t
) -> Option<SBBreakpointLocation>

fn find_location_id_by_address(&self, address: lldb_addr_t) -> i32

fn find_location_by_id(&self, id: i32) -> Option<SBBreakpointLocation>

fn locations(&self) -> SBBreakpointLocationIter