Struct lldb::SBBreakpointLocation

pub struct SBBreakpointLocation {
    pub raw: SBBreakpointLocationRef,
}

One unique instance (by address) of a logical breakpoint.

A breakpoint location is defined by the breakpoint that produces it, and the address that resulted in this particular instantiation. Each breakpoint location has its settable options.

SBBreakpoint contains SBBreakpointLocation(s). See SBBreakpoint for retrieval of an SBBreakpointLocation from an SBBreakpoint.

Fields
raw: SBBreakpointLocationRef

The underlying raw SBBreakpointLocationRef.

Methods

impl SBBreakpointLocation

fn wrap(raw: SBBreakpointLocationRef) -> SBBreakpointLocation

Construct a new SBBreakpointLocation.

fn maybe_wrap(raw: SBBreakpointLocationRef) -> Option<SBBreakpointLocation>

Construct a new Some(SBBreakpointLocation) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBBreakpointLocation value.

fn id(&self) -> i32

fn address(&self) -> Option<SBAddress>

fn load_address(&self) -> lldb_addr_t

fn is_enabled(&self) -> bool

fn set_enabled(&self, enabled: bool)

fn ignore_count(&self) -> u32

fn set_ignore_count(&self, count: u32)

fn is_resolved(&self) -> bool

fn breakpoint(&self) -> SBBreakpoint