Struct lldb::SBFrame

pub struct SBFrame {
    pub raw: SBFrameRef,
}

One of the stack frames associated with a thread.

Fields

raw

The underlying raw SBFrameRef.

Methods

impl SBFrame

fn is_valid(&self) -> bool

Check whether or not this is a valid SBFrame value.

fn frame_id(&self) -> u32

The zero-based stack frame index for this frame.

fn pc_address(&self) -> SBAddress

The program counter (PC) as a section offset address (SBAddress).

fn module(&self) -> SBModule

The SBModule for this stack frame.

fn compile_unit(&self) -> SBCompileUnit

The SBCompileUnit for this stack frame.

fn function(&self) -> SBFunction

The SBFunction for this stack frame.

fn symbol(&self) -> SBSymbol

The SBSymbol for this stack frame.

fn block(&self) -> SBBlock

Get the deepest block that contains the frame PC.

fn function_name(&self) -> &str

Get the appropriate function name for this frame. Inlined functions in LLDB are represented by blocks that have inlined function information, so just looking at the SBFunction or SBSymbol for a frame isn't enough. This function will return the appropriate function, symbol or inlined function name for the frame.

This function returns:

the name of the inlined function (if there is one)

the name of the concrete function (if there is one)

the name of the symbol (if there is one)

NULL

See also is_inlined.

fn display_function_name(&self) -> &str

fn is_inlined(&self) -> bool

Return true if this frame represents an inlined function.

fn frame_block(&self) -> SBBlock

Gets the lexical block that defines the stack frame. Another way to think of this is it will return the block that contains all of the variables for a stack frame. Inlined functions are represented as SBBlock objects that have inlined function information: the name of the inlined function, where it was called from. The block that is returned will be the first block at or above the block for the PC (SBFrame::block()) that defines the scope of the frame. When a function contains no inlined functions, this will be the top most lexical block that defines the function. When a function has inlined functions and the PC is currently in one of those inlined functions, this method will return the inlined block that defines this frame. If the PC isn't currently in an inlined function, the lexical block that defines the function is returned.

fn line_entry(&self) -> SBLineEntry

The line table entry (SBLineEntry) for this stack frame.

fn thread(&self) -> SBThread

The thread that is executing this stack frame.