Struct lldb::SBFrame

pub struct SBFrame {
    pub raw: SBFrameRef,
}

One of the stack frames associated with a thread.

Fields
raw: SBFrameRef

The underlying raw SBFrameRef.

Methods

impl SBFrame

fn wrap(raw: SBFrameRef) -> SBFrame

Construct a new SBFrame.

fn maybe_wrap(raw: SBFrameRef) -> Option<SBFrame>

Construct a new Some(SBFrame) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBFrame value.

fn frame_id(&self) -> u32

The zero-based stack frame index for this frame.

fn cfa(&self) -> Option<lldb_addr_t>

Get the Canonical Frame Address for this stack frame.

This is the DWARF standard's definition of a CFA, a stack address that remains constant throughout the lifetime of the function.

fn pc(&self) -> lldb_addr_t

fn set_pc(&mut self, new_pc: lldb_addr_t) -> bool

fn sp(&self) -> lldb_addr_t

fn fp(&self) -> lldb_addr_t

fn pc_address(&self) -> SBAddress

The program counter (PC) as a section offset address (SBAddress).

fn module(&self) -> SBModule

The SBModule for this stack frame.

fn compile_unit(&self) -> SBCompileUnit

The SBCompileUnit for this stack frame.

fn function(&self) -> SBFunction

The SBFunction for this stack frame.

fn symbol(&self) -> SBSymbol

The SBSymbol for this stack frame.

fn block(&self) -> SBBlock

Get the deepest block that contains the frame PC.

fn function_name(&self) -> Option<&str>

Get the appropriate function name for this frame. Inlined functions in LLDB are represented by blocks that have inlined function information, so just looking at the SBFunction or SBSymbol for a frame isn't enough. This function will return the appropriate function, symbol or inlined function name for the frame.

This function returns:

the name of the inlined function (if there is one)

the name of the concrete function (if there is one)

the name of the symbol (if there is one)

NULL

See also is_inlined.

fn display_function_name(&self) -> Option<&str>

fn is_inlined(&self) -> bool

Return true if this frame represents an inlined function.

fn evaluate_expression(&self, expression: &str, options: &SBExpressionOptions) -> SBValue

Evaluate an expression within the context of this frame.

fn frame_block(&self) -> SBBlock

Gets the lexical block that defines the stack frame. Another way to think of this is it will return the block that contains all of the variables for a stack frame. Inlined functions are represented as SBBlock objects that have inlined function information: the name of the inlined function, where it was called from. The block that is returned will be the first block at or above the block for the PC (SBFrame::block()) that defines the scope of the frame. When a function contains no inlined functions, this will be the top most lexical block that defines the function. When a function has inlined functions and the PC is currently in one of those inlined functions, this method will return the inlined block that defines this frame. If the PC isn't currently in an inlined function, the lexical block that defines the function is returned.

fn line_entry(&self) -> Option<SBLineEntry>

The line table entry (SBLineEntry) for this stack frame.

fn thread(&self) -> SBThread

The thread that is executing this stack frame.

fn disassemble(&self) -> &str

fn variables(&self, options: &SBVariablesOptions) -> SBValueList

fn all_variables(&self) -> SBValueList

fn arguments(&self) -> SBValueList

fn locals(&self) -> SBValueList

fn statics(&self) -> SBValueList

fn registers(&self) -> SBValueList

fn find_register(&self, name: &str) -> Option<SBValue>

fn parent_frame(&self) -> Option<SBFrame>