Struct lldb::SBFunction

pub struct SBFunction {
    pub raw: SBFunctionRef,
}

A generic function, which can be inlined or not.

Fields

raw

The underlying raw SBFunctionRef.

Methods

impl SBFunction

fn wrap(raw: SBFunctionRef) -> SBFunction

Construct a new SBFunction.

fn maybe_wrap(raw: SBFunctionRef) -> Option<SBFunction>

Construct a new Some(SBFunction) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBFunction value.

fn name(&self) -> &str

The name of this function.

fn display_name(&self) -> &str

The display name for the function, as it should be seen in a UI.

fn mangled_name(&self) -> &str

The mangled (linkage) name for this function.

fn start_address(&self) -> SBAddress

Get the address of the start of this function.

fn end_address(&self) -> SBAddress

Get the address of the end of this function.

fn prologue_byte_size(&self) -> u32

Get the size of the function prologue, in bytes.

fn is_optimized(&self) -> bool

Returns true if the function was compiled with optimization.

Optimization, in this case, is meant to indicate that the debugger experience may be confusing for the user -- variables optimized away, stepping jumping between source lines -- and the driver may want to provide some guidance to the user about this. Returns false if unoptimized, or unknown.