Struct lldb::SBInstruction

pub struct SBInstruction {
    pub raw: SBInstructionRef,
}

A machine instruction.

Fields

raw

The underlying raw SBInstructionRef.

Methods

impl SBInstruction

fn wrap(raw: SBInstructionRef) -> SBInstruction

Construct a new SBInstruction.

fn maybe_wrap(raw: SBInstructionRef) -> Option<SBInstruction>

Construct a new Some(SBInstruction) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBInstruction value.

fn address(&self) -> SBAddress

Get the address of the instruction.

fn address_class(&self) -> AddressClass

Get the address class for the address of the instruction.

fn data(&self, target: &SBTarget) -> SBData

fn byte_size(&self) -> u32

fn is_branch(&self) -> bool