Struct lldb::SBProcess

pub struct SBProcess {
    pub raw: SBProcessRef,
}

The process associated with the target program.

Fields

raw

The underlying raw SBProcessRef.

Methods

impl SBProcess

fn wrap(raw: SBProcessRef) -> SBProcess

Construct a new SBProcess.

fn maybe_wrap(raw: SBProcessRef) -> Option<SBProcess>

Construct a new Some(SBProcess) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBProcess value.

fn state(&self) -> StateType

The current state of this process (running, stopped, exited, etc.).

fn is_alive(&self) -> bool

Returns true if the process is currently alive.

fn is_running(&self) -> bool

Returns true if the process is currently running.

fn is_stopped(&self) -> bool

Returns true if the process is currently stopped.

fn exit_status(&self) -> i32

The exit status of the process when the process state is eStateExited.

fn exit_description(&self) -> &str

The exit description of the process when the process state is eStateExited.

fn process_id(&self) -> lldb_pid_t

Returns the process ID of the process.

fn unique_id(&self) -> u32

Returns an integer ID that is guaranteed to be unique across all process instances. This is not the process ID, just a unique integer for comparison and caching purposes.

fn address_byte_size(&self) -> u32

Get the size, in bytes, of an address.

fn destroy(&self) -> Result<(), SBError>

Kills the process and shuts down all threads that were spawned to track and monitor the process.

fn continue_execution(&self) -> Result<(), SBError>

fn stop(&self) -> Result<(), SBError>

fn kill(&self) -> Result<(), SBError>

Same as calling destroy.

fn detach(&self) -> Result<(), SBError>

fn signal(&self, signal: i32) -> Result<(), SBError>

Send the process a Unix signal.

fn broadcaster(&self) -> SBBroadcaster