Struct lldb::SBSymbol

pub struct SBSymbol {
    pub raw: SBSymbolRef,
}

The symbol possibly associated with a stack frame.

Fields
raw: SBSymbolRef

The underlying raw SBSymbolRef.

Methods

impl SBSymbol

fn wrap(raw: SBSymbolRef) -> SBSymbol

Construct a new SBSymbol.

fn maybe_wrap(raw: SBSymbolRef) -> Option<SBSymbol>

Construct a new Some(SBSymbol) or None.

fn is_valid(&self) -> bool

Check whether or not this is a valid SBSymbol value.

fn name(&self) -> &str

The name of this function.

fn display_name(&self) -> &str

The display name for the function, as it should be seen in a UI.

fn mangled_name(&self) -> &str

The mangled (linkage) name for this function.

fn get_instructions(&self, target: &SBTarget, flavor: DisassemblyFlavor) -> SBInstructionList

fn start_address(&self) -> Option<SBAddress>

Get the address that this symbol refers to, if present.

fn end_address(&self) -> Option<SBAddress>

If the symbol has an address and the underlying value has a non-zero size, this will have the address of the end of the value.

Note: It seems unfortunate that if the underlying value is 0-sized, this will result in None rather than the same address as the start_address.

fn prologue_byte_size(&self) -> u32

Get the size of the function prologue, in bytes.

fn symbol_type(&self) -> SymbolType

What type of symbol is this?

fn is_external(&self) -> bool

Is this symbol externally visible (exported) from the module that contains it?

fn is_synthetic(&self) -> bool

Is this symbol synthetically created from information in the module that contains it?