Struct lldb::SBThread

pub struct SBThread {
    pub raw: SBThreadRef,
}

A thread of execution.

SBThreads can be referred to by their ID, which maps to the system specific thread identifier, or by IndexID. The ID may or may not be unique depending on whether the system reuses its thread identifiers. The IndexID is a monotonically increasing identifier that will always uniquely reference a particular thread, and when that thread goes away it will not be reused.

Fields

raw

The underlying raw SBThreadRef.

Methods

impl SBThread

fn is_valid(&self) -> bool

Check whether or not this is a valid SBThread value.

fn stop_reason(&self) -> LLDBStopReason

Get the stop reason for this thread.

fn stop_return_value(&self) -> SBValue

The return value from the last stop if we just stopped due to stepping out of a function

fn thread_id(&self) -> lldb_tid_t

Returns a unique thread identifier for the current SBThread that will remain constant throughout the thread's lifetime in this process and will not be reused by another thread during this process lifetime. On Mac OS X systems, this is a system-wide unique thread identifier; this identifier is also used by other tools like sample which helps to associate data from those tools with lldb. See related SBThread::index_id.

fn index_id(&self) -> u32

Return the index number for this SBThread. The index number is the same thing that a user gives as an argument to thread select in the command line lldb.

These numbers start at 1 (for the first thread lldb sees in a debug session) and increments up throughout the process lifetime. An index number will not be reused for a different thread later in a process - thread 1 will always be associated with the same thread. See related SBThread::thread_id.